+++ example/lib/page/page_five.dart
@@ -106,7 +106,7 @@ class _PageFiveState extends State<PageFive> {
     DLog.d("result: $result");
   }
 
-  void showSheet() {
+  Future<void> showSheet() async {
     showModalBottomSheet(
       context: context,
       builder: (context) {