+++ lib/src/route_manager.dart
@@ -7,9 +7,8 @@
 //
 
 import 'dart:convert';
-
 import 'dart:developer' as developer;
-
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 /// 路由堆栈管理器
@@ -22,14 +21,23 @@ class RouteManager {
   /// 是否打印日志
   bool isDebug = false;
 
-  /// 路由堆栈(满足 filterRoute 的)
-  final List<Route<dynamic>> _routes = [];
-
   /// 所有路由堆栈
-  final List<Route<dynamic>> _allRoutes = [];
+  final List<Route<dynamic>> _routes = [];
 
   /// 当前路由堆栈
   List<Route<dynamic>> get routes => _routes;
+
+  /// 当前 PageRoute 路由堆栈
+  List<PageRoute<dynamic>> get pageRoutes =>
+      _routes.whereType<PageRoute>().toList();
+
+  /// 当前 DialogRoute 路由堆栈
+  List<DialogRoute<dynamic>> get dialogRoutes =>
+      _routes.whereType<DialogRoute>().toList();
+
+  /// 当前 ModalBottomSheetRoute 路由堆栈
+  List<ModalBottomSheetRoute<dynamic>> get sheetRoutes =>
+      _routes.whereType<ModalBottomSheetRoute>().toList();
 
   /// 当前路由名堆栈
   List<String?> get routeNames => routes.map((e) => e.settings.name).toList();
@@ -37,26 +45,19 @@ class RouteManager {
   /// 之前路由
   Route<dynamic>? preRoute;
 
-  /// 当前路由
+  /// 之前路由 name
   String? get preRouteName => preRoute?.settings.name;
 
   /// 当前路由
   Route<dynamic>? get currentRoute => routes.isEmpty ? null : routes.last;
 
-  /// 当前路由
-  String? get current => currentRoute?.settings.name;
+  /// 当前路由 name
+  String? get currentRouteName => currentRoute?.settings.name;
 
-  /// PopupRoute 类型路由
+  /// 最近的 PopupRoute 类型路由
   PopupRoute? get popupRoute {
-    // for (int i = routes.length - 1; i >= 0; i--) {
-    //   final e = routes[i];
-    //   if (e is PopupRoute) {
-    //     return e;
-    //   }
-    // }
-
-    for (int i = _allRoutes.length - 1; i >= 0; i--) {
-      final e = _allRoutes[i];
+    for (int i = routes.length - 1; i >= 0; i--) {
+      final e = routes[i];
       if (e is PopupRoute) {
         return e;
       }
@@ -73,9 +74,9 @@ class RouteManager {
   /// 路由堆栈包含 ModalBottomSheetRoute 类型
   bool get isSheetOpen => popupRoute is ModalBottomSheetRoute;
 
-  /// 进出堆栈过滤条件(默认仅支持PageRoute, 过滤弹窗)
-  bool Function(Route<dynamic> route) filterRoute =
-      (route) => route is PageRoute && route.settings.name != null;
+  // /// 进出堆栈过滤条件(默认仅支持PageRoute, 过滤弹窗)
+  // bool Function(Route<dynamic> route) filterRoute =
+  //     (route) => route is PageRoute && route.settings.name != null;
 
   /// 更新回调
   ValueChanged<RouteManager>? onChanged;
@@ -87,44 +88,40 @@ class RouteManager {
 
   /// 路由对应的参数
   Object? getArguments(String routeName) {
-    final route = routes.firstWhere((e) => e.settings.name == routeName);
-    return route.settings.arguments;
+    final index = pageRoutes.indexWhere((e) => e.settings.name == routeName);
+    if (index == -1) {
+      return null;
+    }
+    final route = pageRoutes[index];
+    return route;
   }
 
   /// 入栈
   void push(Route<dynamic> route) {
-    if (_allRoutes.isEmpty ||
-        _allRoutes.isNotEmpty && _allRoutes.last != route) {
-      _allRoutes.add(route);
+    if (_routes.isEmpty || _routes.isNotEmpty && _routes.last != route) {
+      _routes.add(route);
     }
-
-    if (!filterRoute(route)) {
-      debugPrint("❌push ${[route.runtimeType, route.settings.name]}");
-      return;
-    }
-    if (_routes.isNotEmpty &&
-        _routes.last.settings.name == route.settings.name) {
-      return;
-    }
-    _routes.add(route);
   }
 
   /// 出栈
   void pop(Route<dynamic> route) {
-    _allRoutes.remove(route);
-    if (!filterRoute(route)) {
-      return;
-    }
-    _routes.removeWhere((e) => e.settings.name == route.settings.name);
+    _routes.remove(route);
   }
 
   Map<String, dynamic> toJson() {
     final data = <String, dynamic>{};
     data['isDebug'] = isDebug;
     data['routes'] = routes.map((e) => e.toString()).toList();
+    data['pageRoutes'] = pageRoutes.map((e) => e.toString()).toList();
+    if (dialogRoutes.isNotEmpty) {
+      data['dialogRoutes'] = dialogRoutes.map((e) => e.toString()).toList();
+    }
+    if (sheetRoutes.isNotEmpty) {
+      data['sheetRoutes'] = sheetRoutes.map((e) => e.toString()).toList();
+    }
     data['routeNames'] = routeNames;
     data['preRouteName'] = preRouteName;
-    data['current'] = current;
+    data['currentRouteName'] = currentRouteName;
     data['popupRoute'] = popupRoute.toString();
     data['isPopupOpen'] = isPopupOpen;
     data['isDialogOpen'] = isDialogOpen;
@@ -145,6 +142,7 @@ class RouteManager {
     if (!isDebug) {
       return;
     }
+
     developer.log(toString());
   }
 }